chore(structure): move GeoParquetDuckDB frontend files up one level and fix imports
