+++ src/maps/GeoParquetDuckDB_2D/components/HUD.jsx
