+++ src/maps/GeoParquetDuckDB_2D/components/PlaybackControls.jsx
