+++ src/maps/GeoParquetDuckDB_3D/components/PlaybackControls.jsx
