+++ src/maps/config.js
@@ -13,8 +13,8 @@ import geoParquetDemo from './GeoParquetDemo/index.jsx';
 import geoParquetLoaders from './GeoparquetLoaders/index.jsx';
 import geoParquetLoadersBinary from './GeoparquetLoadersBinary/index.jsx';
 import geoParquetDuckDB from './GeoParquetDuckDB/index.jsx';
-import geoParquetDuckDB3D from './GeoParquetDuckDB_3D/frontend/Map3D.jsx';
-import geoParquetDuckDB2D from './GeoParquetDuckDB_2D/frontend/Map2D.jsx';
+import geoParquetDuckDB3D from './GeoParquetDuckDB_3D/Map3D.jsx';
+import geoParquetDuckDB2D from './GeoParquetDuckDB_2D/Map2D.jsx';
 import geoParquetDuckDBwasm from './GeoParquetDuckDBwasm/MinimalWasmMap.jsx';
 import geoParquetTiled from './GeoParquetTiled/GeoParquetTiled.jsx';
 import geoParquetVirtualTiles from './GeoParquetVirtualTile/VirtualTilesMap.jsx';
